5.00.20
Some updates for ArcGis pro

+++ Toolbox/arcsdm/calculateweights.py
@@ -207,7 +207,7 @@ def Calculate(self, parameters, messages):
         Confident_Contrast = float( parameters[5].valueAsText)
         #Unitarea = float( parameters[6].valueAsText)
         Unitarea = float( parameters[6].value)
-        MissingDataValue = long( parameters[7].valueAsText)
+        MissingDataValue = int( parameters[7].valueAsText) # Python 3 fix, long -> int
         #gp.AddMessage("Debug step 12");
         arcsdm.sdmvalues.appendSDMValues(gp,  Unitarea, TrainingSites)
         arcpy.AddMessage("="*10 + " Calculate weights " + "="*10)
